+++ lib/gem_updater/ruby_gems_fetcher.rb
@@ -8,23 +8,22 @@ module GemUpdater
   # RubyGemsFetcher is a wrapper around rubygems API.
   class RubyGemsFetcher
     HTTP_TOO_MANY_REQUESTS = '429'
-    GEM_HOMEPAGES = %w[source_code_uri homepage_uri].freeze
 
-    attr_reader :gem_name, :source
+    attr_reader :gem_name
 
     # @param gem_name [String] name of the gem
-    # @param source [Bundler::Source] source of gem
-    def initialize(gem_name, source)
+    def initialize(gem_name)
       @gem_name = gem_name
-      @source   = source
     end
 
-    # Finds where code is hosted.
-    # Most likely in will be on rubygems, else look in other sources.
+    # Finds the changelog uri.
+    # It asks rubygems.org for changelog_uri of gem.
+    # See API: http://guides.rubygems.org/rubygems-org-api/#gem-methods
     #
-    # @return [String|nil] url of gem source code
-    def source_uri
-      uri_from_rubygems || uri_from_other_sources
+    # @return [String|nil] uri of changelog
+    def changelog_uri
+      response = query_rubygems
+      response.to_h['changelog_uri']
     end
 
     private
@@ -35,19 +34,6 @@ module GemUpdater
     # @return [Hash] parsed JSON
     def parse_remote_json(url)
       JSON.parse(URI.parse(url).open.read)
-    end
-
-    # Ask rubygems.org for source uri of gem.
-    # See API: http://guides.rubygems.org/rubygems-org-api/#gem-methods
-    #
-    # @return [String|nil] uri of source code
-    def uri_from_rubygems
-      return unless source.remotes.map(&:host).include?('rubygems.org')
-
-      response = query_rubygems
-      return unless response
-
-      response[GEM_HOMEPAGES.find { |key| response[key] && !response[key].empty? }]
     end
 
     # Make the real query to rubygems
@@ -63,43 +49,5 @@ module GemUpdater
         sleep 1 && retry if tries < 2
       end
     end
-
-    # Look if gem can be found in another remote
-    #
-    # @return [String|nil] uri of source code
-    def uri_from_other_sources
-      source.remotes.find do |remote|
-        case remote.host
-        when 'rubygems.org' then next # already checked
-        when 'rails-assets.org'
-          return uri_from_railsassets
-        else
-          Bundler.ui.error "Source #{remote} is not supported, ' \
-            'feel free to open a PR or an issue on https://github.com/MaximeD/gem_updater"
-        end
-      end
-    end
-
-    # Ask rails-assets.org for source uri of gem.
-    # API is at : https://rails-assets.org/packages/package_name
-    #
-    # @return [String|nil] uri of source code
-    def uri_from_railsassets
-      response = query_railsassets
-      return unless response
-
-      response['url'].gsub(/^git/, 'http')
-    end
-
-    # Make the real query to railsassets
-    # rubocop:disable Lint/SuppressedException
-    def query_railsassets
-      parse_remote_json("https://rails-assets.org/packages/#{gem_name.gsub('rails-assets-', '')}")
-    rescue JSON::ParserError
-      # if gem is not found, rails-assets returns a 200
-      # with html (instead of json) containing a 500...
-    rescue OpenURI::HTTPError
-    end
-    # rubocop:enable Lint/SuppressedException
   end
 end